+++ api/src/controllers/health.controller.ts
@@ -1,5 +1,6 @@
 import { Controller, Get } from '@nestjs/common';
 import { ApiUseTags, ApiExcludeEndpoint } from '@nestjs/swagger';
+import { version } from '../../package.json';
 
 @Controller()
 export default class HealthController {
@@ -8,6 +9,6 @@ export default class HealthController {
   @ApiExcludeEndpoint()
   @Get('/health')
   async health() {
-    return { status: 'ok' };
+    return { status: 'ok', version };
   }
 }